.emacs: take display dimensions via chj-rustbin's `e`
Because it's a hen-and-egg issue: Emacs won't report the x-display
dimensions before it has opened a frame, but we want the dimensions to
calculate the frame size. So `e` will set
`chj-rustbin:x-display-dimensions` before opening a frame if it just
started the daemon.

PS.: after coding up the Rust part, I realize that this change in the
elisp code alone now made it work already (no setting from `e`
needed). Why??

Because I'm *not* setting geometry defaults anymore, does emacs "do it
fresh" then? But why is the size calculation mine?